move optional model field tests into single file
